Reorganise ops, processors and utils as directories
