+++ Wrappers/Python/wip/demo_astra_mc.py
@@ -8,7 +8,7 @@
 from ccpi.framework import ImageData, AcquisitionData, ImageGeometry, AcquisitionGeometry
 from ccpi.optimisation.algs import FISTA
 from ccpi.optimisation.funcs import Norm2sq, Norm1
-from ccpi.astra.ops import AstraProjectorMC
+from ccpi.astra.operators import AstraProjectorMC
 
 import numpy
 import matplotlib.pyplot as plt
